+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+
+### Bug fixes
+
+- [#100][GL-100] - ArrayIndexOutOfBoundsException when using the cave-ores-only setting
+
+[GL-100]: https://gitlab.com/uhccore/uhccore/-/issues/100
+
 ## [1.20.6] - 2023-08-05
 
 UhcCore version `1.20.6` is now here and brings a whole bunch of bug fixes and new features!